requirements.txt
requests>=2.32.3
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
pytest>=8.0.0
Flask>=3.0.3
vaderSentiment>=3.3.2
spacy>=3.7.2
python-dotenv>=1.0.1
